models/backbone_model.py
import torch
import torch.nn as nn


class CustomObjectDetector(nn.Module):
    def __init__(self, num_classes):
        super(CustomObjectDetector, self).__init__()
        # Define the backbone (feature extractor), here we use a simple CNN
        self.backbone = nn.Sequential(
            nn.Conv2d(3, 32, kernel_size=3, stride=1, padding=1),
            nn.BatchNorm2d(32),
            nn.ReLU(inplace=True),
            nn.MaxPool2d(kernel_size=2, stride=2),
            # ... (more layers can be added)
        )

        # Define the RPN (Region Proposal Network)
        self.rpn = self._make_rpn_layer()

        # Define the classifier head
        self.classifier_head = self._make_head_layer(num_classes)

        # Define the regressor head for bounding box regression
        self.regressor_head = self._make_head_layer(
            4)  # 4 coordinates for bbox

    def forward(self, x):
        # Feature extraction
        features = self.backbone(x)

        # Generate proposals with RPN (dummy implementation)
        proposals = self.rpn(features)

        # Predict classes and bbox deltas for each proposal (dummy implementation)
        class_logits = self.classifier_head(features)
        bbox_deltas = self.regressor_head(features)

        return proposals, class_logits, bbox_deltas

    def _make_rpn_layer(self):
        # Dummy RPN layer
        return nn.Sequential(
            nn.Conv2d(32, 128, kernel_size=3, stride=1, padding=1),
            nn.ReLU(inplace=True),
            # Normally you would have anchor generation and proposal calculation here
        )

    def _make_head_layer(self, output_size):
        # Dummy head layer
        return nn.Sequential(
            nn.Flatten(),
            nn.Linear(32 * 64 * 64, 512),
            nn.ReLU(inplace=True),
            nn.Linear(512, output_size),
        )


# # Example usage:
# num_classes = 10  # Example number of classes
# model = CustomObjectDetector(num_classes=num_classes)
#
# # Print the model to see its architecture
# print(model)

# Continue with model training and evaluation as appropriate...
